backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pytesseract==0.3.10
Pillow>=10.4.0
python-multipart==0.0.9
pandas>=2.2.2
scikit-learn>=1.4.2
numpy>=1.26.4
aiofiles==23.2.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
anyio==4.3.0
pdf2image
openai-whisper
torch
more-itertools
setuptools-rust
ffmpeg-python